fix databaseSlice
